pyproject.toml
[tool.poetry]
name = "telegram-proxy-bot"
version = "0.1.0"
description = "A Telegram bot to collect and send proxies and configs with admin dashboard"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
geoip2 = "^4.8.0"
dnspython = "^2.7.0"
apscheduler = "^3.10.4"
python-dotenv = "^1.1.1"
pytelegrambotapi = "^4.28.0"
supabase = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"